# packages/contracts/openapi/runtime-gateway.openapi.yaml
openapi: 3.1.0
info:
  title: AgentTwin runtime gateway API
  version: 1.0.0
  summary: Runtime containment of an agent's tool actions — policies, approvals and every decision.
  description: |
    The runtime gateway is an opt-in path through which an agent's tool
    actions pass at runtime (spec §30-31, ADR-0033). Each call is decided by
    the tool's active policies: allowed, allowed with limits, sent to a
    person for approval, or denied — before it reaches the tool.

    * **Tool endpoints** name the tools the gateway may call, their risk tier
      and timeout; a URL outside the egress allowlist is refused.
    * **Policies** are named, versioned documents of CEL rules that guard one
      tool. A version is tested (its own tests and every numeric threshold)
      and activated only when every test passes.
    * **Approvals** bind a person's decision to one exact action: the approved
      action runs once; a changed action needs its own approval.
    * **Decisions** record every call the gateway decided (append-only).

    **Where to call it.** The public API is served by the control plane (the
    edge), which authenticates the caller, applies rate limits and, on
    `/api/v1` mutations, `Idempotency-Key`; it forwards the request to the
    runtime gateway. The invocation path (`/gateway/v1`) is reached the same
    way with an API key holding the `runtime:invoke` scope; there the
    gateway itself implements `Idempotency-Key`, per action, and forwards the
...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /api/v1/policies:
    get:
      operationId: listPolicies
      tags: [Policies]
      summary: List policies
      description: The project's policies by name, optionally those guarding one tool.
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
        - name: tool
          in: query
          schema:
            $ref: '#/components/schemas/ToolName'
      responses:
        '200':
          description: The policies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
    post:
      operationId: createPolicy
      tags: [Policies]
      summary: Create a policy
      description: |
        Stores a policy document (YAML or JSON, `apiVersion:
        agenttwin.dev/v1`, `kind: Policy`) as version 1 of a new policy named
        by its `metadata.name`, guarding `spec.tool`. Every rule is compiled
        and type-checked; nothing is activated. Needs `policy.write`; audited
        (`policy.created`).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyDocumentInput'
      responses:
        '201':
          description: The policy and its first version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyCreated'
        '400':
          $ref: '#/components/responses/PolicyInvalid'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/PolicyExists'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/policies/test:
    post:
      operationId: testPolicy
      tags: [Policies]
      summary: Test a policy
      description: |
        Decides a draft document, or a saved version, against its own tests
        and the extra `cases`, and around every numeric threshold its rules
        compare against: `args.amount > 100` is decided at 99, 99.99, 100,
        100.01 and 101 (one hundredth either side of a decimal threshold; one
        call either side of `trace.calls`). The boundaries start from `base`,
        else from the first test's action. Reports the rules no test decided
        and whether the version could be activated for the tool as
        registered. Needs `policy.test`; nothing is stored.
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyTestInput'
      responses:
        '200':
          description: What the policy decided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyTestReport'
        '400':
          $ref: '#/components/responses/PolicyTestInvalid'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyInProgress'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/policies/{policy_id}:
    parameters:
      - $ref: '#/components/parameters/PolicyId'
    get:
      operationId: getPolicy
      tags: [Policies]
      summary: Read a policy
      description: The policy, its active version and every version (newest first).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
      responses:
        '200':
          description: The policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A session token (dev login), an OIDC access token or an API key.
    apiKey:
      type: apiKey
      in: header
      name: X-AgentTwin-Api-Key
      description: A project API key.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        One key per user action (ADR-0019), handled by the edge. A repeat of
        the same request returns the stored response with
        `Idempotent-Replayed: true`; the same key with a different request
        answers `422`; while the first request runs, `409`. Keys expire after
        24 hours.
      schema:
        type: string
        pattern: ^[A-Za-z0-9._:-]{8,128}$
    Organization:
      name: X-AgentTwin-Org
      in: header
      required: false
      description: The organization to act in, for credentials that belong to several.
      schema:
        $ref: '#/components/schemas/Uuid'
    Project:
      name: project_id
      in: query
      required: true
      description: The project. A project outside the caller's access answers `404`.
      schema:
        $ref: '#/components/schemas/Uuid'
...
    PolicyId:
      name: policy_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
    BadRequest:
      description: |
        The request is invalid: `INVALID_PARAMETER` (`details.field`),
        `INVALID_CURSOR`, `PROJECT_REQUIRED` or `INVALID_IDEMPOTENCY_KEY`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    PolicyInvalid:
      description: |
        `INVALID_POLICY` (`details.problems`: every problem with its field —
        schema, names, rules that do not compile or are not boolean, tests;
        a new version that renames the policy or retargets its tool),
        `INVALID_PARAMETER`, `INVALID_JSON`, `EMPTY_BODY` or
        `INVALID_IDEMPOTENCY_KEY`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PolicyTestInvalid:
      description: |
        `INVALID_POLICY` (a draft document that is not valid;
        `details.problems`), `INVALID_CASES` (`details.field`),
        `INVALID_PARAMETER` (neither or both of a document and a saved
        version), `INVALID_JSON`, `EMPTY_BODY` or `INVALID_IDEMPOTENCY_KEY`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Unauthenticated:
      description: No valid credential (`UNAUTHENTICATED`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The credential lacks the permission (`FORBIDDEN`; `details.required_permission`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The project or resource does not exist or is outside the caller's projects (`NOT_FOUND`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PolicyExists:
      description: |
        A policy of that name exists in the project (`POLICY_EXISTS`; add a
        version to it), or the first request with this `Idempotency-Key` is
        still running (`IDEMPOTENCY_IN_PROGRESS`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    IdempotencyInProgress:
      description: The first request with this `Idempotency-Key` is still running (`IDEMPOTENCY_IN_PROGRESS`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: The request body is too large (`PAYLOAD_TOO_LARGE`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnsupportedMediaType:
      description: The body is not `application/json` (`UNSUPPORTED_MEDIA_TYPE`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    IdempotencyKeyReused:
      description: The `Idempotency-Key` was used for a different request (`IDEMPOTENCY_KEY_REUSED`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Too many requests (`RATE_LIMITED`).
      headers:
        Retry-After:
          description: Seconds to wait.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Internal:
      description: An unexpected error (`INTERNAL`); quote the `request_id`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UpstreamUnavailable:
      description: The runtime gateway is not reachable from the control plane (`UPSTREAM_UNAVAILABLE`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unavailable:
      description: |
        The runtime gateway is not configured on this control plane
        (`SERVICE_NOT_CONFIGURED`) or its database is unavailable
        (`UNAVAILABLE`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ToolName:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$
...
    Hash:
      type: string
      pattern: ^[0-9a-f]{64}$
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
...
    Policy:
      type: object
      required: [id, name, tool, description, latest_version, active, activated_by, activated_at, created_by, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        tool:
          $ref: '#/components/schemas/ToolName'
        description:
          type: string
        latest_version:
          type: integer
        active:
          description: The active version, or `null` when the policy guards nothing.
          oneOf:
            - $ref: '#/components/schemas/VersionRef'
            - type: 'null'
        activated_by:
          type: [string, 'null']
        activated_at:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        created_by:
          type: string
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    PolicyPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
...
    PolicyDetail:
      allOf:
        - $ref: '#/components/schemas/Policy'
        - type: object
          required: [versions]
          properties:
            versions:
              type: array
              items:
                $ref: '#/components/schemas/VersionSummary'
...
    PolicyDocumentInput:
      type: object
      additionalProperties: false
      required: [project_id, document]
      properties:
        project_id:
          $ref: '#/components/schemas/Uuid'
        document:
          type: string
          maxLength: 262144
          description: The policy document, YAML or JSON.
    PolicyCreated:
      type: object
      required: [policy, version]
      properties:
        policy:
          $ref: '#/components/schemas/Policy'
        version:
          $ref: '#/components/schemas/PolicyVersion'
...
    PolicyTestInput:
      type: object
      additionalProperties: false
      required: [project_id]
      properties:
        project_id:
          $ref: '#/components/schemas/Uuid'
        document:
          type: string
          maxLength: 262144
          description: A draft document (or give `policy_id` and `version`).
        policy_id:
          $ref: '#/components/schemas/Uuid'
        version:
          type: integer
          minimum: 1
        cases:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/PolicyTestCase'
        base:
          type: object
          additionalProperties: false
          description: The action the boundaries start from.
          properties:
            args:
              type: object
            context:
              $ref: '#/components/schemas/TestContext'
...
    PolicyTestReport:
      type: object
      required: [tool, spec_hash, passed, results, boundaries, undecided, tool_risk, activatable, activation_problems]
      properties:
        tool:
          $ref: '#/components/schemas/ToolName'
        spec_hash:
          $ref: '#/components/schemas/Hash'
        passed:
          type: boolean
        results:
          type: array
          items:
            $ref: '#/components/schemas/TestResult'
        boundaries:
          type: array
          items:
            $ref: '#/components/schemas/Boundary'
        undecided:
          type: array
          description: Rules no test decided.
          items:
            type: string
        tool_risk:
          description: The tool's risk tier as registered, `null` when it is not.
          oneOf:
            - $ref: '#/components/schemas/Risk'
            - type: 'null'
        activatable:
          type: boolean
        activation_problems:
          type: array
          items:
            $ref: '#/components/schemas/Problem'
...